Refactor

+++ backend/src/main/java/com/skywalker/backend/repository/MilestoneRepository.java
@@ -3,5 +3,5 @@ package com.skywalker.backend.repository;
 import com.skywalker.backend.model.Milestone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MilestoneRepository extends JpaRepository<Milestone,Integer> {
+public interface MilestoneRepository extends JpaRepository<Milestone,Long> {
 }
